Put an end to the suffering of ItemDetailsActivity
Rewrite the code of ItemDetailsActivity in order to throw out that old junk of a code. This commit also is like a continuation to the previous one as it modifies the API files too: moves some classes to their own files, implements Serializable for some missed classes... Another thing that changes with this commit is that when starting ItemDetailsActivity, we no longer need to pass many args. Just passing the Item is enough. Depending on what the item is, ItemDetailsActivity will either show data directly off the Item (when item is complete), or fetch a complete item (when item is incomplete. e.g., UnlockItem or CriticalItem.)